查看同一事务的读写是否分离

+++ hello-server/hello-server-entity/src/main/java/me/sta/dao/TestMapper.java
@@ -15,5 +15,5 @@ public interface TestMapper {
 
     void deleteByPrimaryKey(Integer id);
 
-    Test findByUserIdAndId(int userId, int id);
+    Test findByUserIdAndId(int userId, Long id);
 }
