+++ hello-server/hello-server-provider-sharding/src/main/java/me/sta/service/TestService.java
@@ -9,5 +9,5 @@ public interface TestService {
 
     int save(String name,Integer userId);
 
-    Test findByUserIdAndId(int userId, int id);
+    Test findByUserIdAndId(int userId, Long id);
 }
